+++ components/display/channel/channel-header.tsx
@@ -45,6 +45,7 @@ export const ChannelHeader = ({
       />
       <RightChannelToggle 
         anotherMemberProp={membersList}
+        userProfileIdProp={userProfileIdProp}
       />
     </div>
   )