+++ components/display/friend-profile.tsx
@@ -3,7 +3,7 @@
 import { usePrompt } from "@/hooks/use-prompt-store";
 import { UserProfileAvatar } from "../uihelper/user-profile-avatar";
 import { FriendWithProfile } from "@/type";
-import { DropdownMenu, DropdownMenuContent, DropdownMenuItem, DropdownMenuSeparator, DropdownMenuSub, DropdownMenuSubContent, DropdownMenuSubTrigger, DropdownMenuTrigger } from "../ui/dropdown-menu";
+import { DropdownMenu, DropdownMenuContent, DropdownMenuItem, DropdownMenuTrigger } from "../ui/dropdown-menu";
 import { Cross } from "lucide-react";
 import { useRouter } from "next/navigation";
 import qs from "query-string"
@@ -26,14 +26,15 @@ const FriendProfileComponent = (
     const {onOpen} = usePrompt();
     const router = useRouter();
 
-    const onUserStatusChange = async(friendRequestId: string,friendConfirm: FriendRequestStatus) => {
+    const onFriendChange = async(friendRequestId: string,friendConfirm: FriendRequestStatus) => {
       try {
           const url = qs.stringifyUrl({
               url: `/api/friend/confirm/${friendRequestId}`,
           })
 
+          console.log("friend confirm status check",friendConfirm)
           const response = await axios.patch(url, {friendConfirm});
-
+          console.log("Friend request respond",response)
           router.refresh();
       }
       catch(error)
@@ -57,7 +58,7 @@ const FriendProfileComponent = (
                     className={
                       "font-semibold text-sm flex flex-row text-zinc-500 group-hover:text-zinc-600 dark:text-zinc-400 dark:group-hover:text-zinc-300 transition "
                     }
-                      onClick={()=>onOpen("UserProfile", {userProfilePropAPI:whichFriend, currentUserPropAPIID: profileId },)}
+                      onClick={()=>onOpen("UserProfileDisplay", {userProfilePropAPI:whichFriend, currentUserPropAPIID: profileId },)}
                     >
                       <UserProfileAvatar 
                         src={whichFriend.imageUrl}
@@ -91,12 +92,12 @@ const FriendProfileComponent = (
                         </DropdownMenuTrigger>
                         <DropdownMenuContent>
                           <DropdownMenuItem
-                            onClick={() => {onUserStatusChange(friendMember.id,FriendRequestStatus.ACCEPT)}}
+                            onClick={() => {onFriendChange(friendMember.id,FriendRequestStatus.ACCEPT)}}
                           >
                             Accept
                           </DropdownMenuItem>
                           <DropdownMenuItem
-                            onClick={() => {onUserStatusChange(friendMember.id,FriendRequestStatus.REFUSE)}}
+                            onClick={() => {onFriendChange(friendMember.id,FriendRequestStatus.REFUSE)}}
                           >
                             Refuse
                           </DropdownMenuItem>